Create README.md
Ongoing Expectations
Going Live
The supplier will manage the transition to the new site at an agreed time & date. The supplier will implement a considered 301 redirect mapping of pages from the old site to the new, and rectify any gaps that might be apparent on having moved to the new site.
Ongoing maintenance / support
We are interested in an ongoing package to update the site, maintain security and fix any new bugs.
Training
We would like to book in half a day of training
Ongoing communication
We expect regular email communication and a weekly stand up meeting to go through progress.
Replying to Spec
Our choice criteria
Understanding of the project
Proven expertise in retail website development and e-commerce
Familiarity with the UK retail sector
Extensive and current experience in providing web hosting and support services
Extensive and current experience in working effectively with small and medium sized retail businesses
Effective project and business planning skills as demonstrated by a realistic and viable initial outline budget for the project

Case Studies

Please include background material to demonstrate the previous experience and range of work of the consultant(s) and of the specific personnel proposed for the work. This can include descriptions and examples of previous relevant work in the form of case studies.

Please provide a breakdown of the days work involved and your day rate