Create CONTRIBUTING.md
Closing notes
Leave a few claps and a star in the project repository if you found this interesting.

I’m indebted to Quincy Larson for advising me to write this article and all his work for freeCodeCamp. Also Kent C. Dodds and Scott Hanselman for inspiring me to start this project.